+++ user_payments/exceptions.py
@@ -1,0 +1,4 @@
+class UnknownPeriodicity(Exception):
+    """The periodicity is unknown"""
+
+    pass